remove Oval
