+++ bokehjs/src/lib/models/glyphs/index.ts
@@ -4,7 +4,6 @@ export {Arc}             from "./arc"
 export {Bezier}          from "./bezier"
 export {Circle}          from "./circle"
 export {Ellipse}         from "./ellipse"
-export {EllipseOval}     from "./ellipse_oval"
 export {Glyph}           from "./glyph"
 export {HArea}           from "./harea"
 export {HBar}            from "./hbar"
@@ -15,7 +14,6 @@ export {ImageURL}        from "./image_url"
 export {Line}            from "./line"
 export {MultiLine}       from "./multi_line"
 export {MultiPolygons}   from "./multi_polygons"
-export {Oval}            from "./oval"
 export {Patch}           from "./patch"
 export {Patches}         from "./patches"
 export {Quad}            from "./quad"
