--- sphinx/source/docs/reference/models/glyphs/oval.rst
@@ -1,4 +1,0 @@
-.. _bokeh.models.glyphs.Oval:
-
-.. bokeh-model:: Oval
-    :module: bokeh.models.glyphs
